+++ src/main/java/io/github/astrapi69/gradle/migration/extension/DependenciesExtensions.java
@@ -40,7 +40,7 @@ import io.github.astrapi69.collection.list.ListFactory;
 import io.github.astrapi69.collection.map.MapFactory;
 import io.github.astrapi69.collection.properties.PropertiesExtensions;
 import io.github.astrapi69.file.read.ReadFileExtensions;
-import io.github.astrapi69.gradle.migration.data.DependencyInfo;
+import io.github.astrapi69.gradle.migration.info.DependencyInfo;
 import io.github.astrapi69.gradle.migration.runner.GradleRunConfigurationsCopier;
 import io.github.astrapi69.string.StringExtensions;
 import lombok.NonNull;