+++ src/main/java/io/github/astrapi69/gradle/migration/info/DependencyInfo.java
@@ -22,27 +22,20 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.gradle.migration.data;
+package io.github.astrapi69.gradle.migration.info;
 
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.meanbean.test.BeanTester;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Data;
+import lombok.experimental.FieldDefaults;
 
-/**
- * The unit test class for the class {@link DependenciesInfo}
- */
-public class DependenciesInfoTest
+@Data
+@Builder
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class DependencyInfo
 {
-
-	/**
-	 * Test method for {@link DependenciesInfo}
-	 */
-	@Test
-	@Disabled
-	public void testWithBeanTester()
-	{
-		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(DependenciesInfo.class);
-	}
-
+	String scope;
+	String groupId;
+	String artifactId;
+	String version;
 }