+++ src/main/java/io/github/astrapi69/gradle/migration/toml/MigrateToTomlVersions.java
@@ -1,8 +1,30 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2023 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.gradle.migration.toml;
 
 import static io.github.astrapi69.gradle.migration.extension.DependenciesExtensions.getLibsVersionTomlMapAsString;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
@@ -10,32 +32,22 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
 
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
 import io.github.astrapi69.file.copy.CopyFileExtensions;
 import io.github.astrapi69.file.create.FileFactory;
-import io.github.astrapi69.file.read.ReadFileExtensions;
 import io.github.astrapi69.file.search.PathFinder;
 import io.github.astrapi69.file.write.StoreFileExtensions;
-import io.github.astrapi69.gradle.migration.data.DependenciesInfo;
-import io.github.astrapi69.gradle.migration.data.DependencyInfo;
 import io.github.astrapi69.gradle.migration.extension.DependenciesExtensions;
+import io.github.astrapi69.gradle.migration.info.DependencyInfo;
+import io.github.astrapi69.gradle.migration.info.MigrationInfo;
+import io.github.astrapi69.gradle.migration.info.ProjectTomlStructureInfo;
 import io.github.astrapi69.gradle.migration.runner.GradleRunConfigurationsCopier;
 
 /**
- * The unit test class for migrate to toml versions
+ * The class {@link MigrateToTomlVersions} provides methods for migrate to new toml project
+ * structure
  */
 public class MigrateToTomlVersions
 {
-
-	public static File getGradleDirectory() throws IOException
-	{
-		File projectDirectory = PathFinder.getProjectDirectory();
-		File gradleDirectory = PathFinder.getRelativePath(projectDirectory, "gradle");
-		return gradleDirectory;
-	}
-
 
 	public static String newLibsVersionsTomlAsString(MigrationInfo migrationInfo) throws IOException
 	{
@@ -44,7 +56,6 @@ public class MigrateToTomlVersions
 
 		File gradlePropertiesFile = PathFinder.getRelativePath(migrationInfo.getProjectDirectory(),
 			"gradle.properties");
-		assertTrue(gradlePropertiesFile.exists());
 		// 1. Load all version from gradle.properties
 		Map<String, String> versionMap = DependenciesExtensions.getVersionMap(gradlePropertiesFile,
 			"Version");
@@ -102,55 +113,6 @@ public class MigrateToTomlVersions
 		return libsVersionsTomlFile;
 	}
 
-	@Test
-	@Disabled
-	public void testNewLibsVersionsTomlFile() throws IOException
-	{
-		File gradleDirectory = getGradleDirectory();
-
-		String targetProjectName = "xstream-extensions";
-		String targetProjectDirNamePrefix = "/run/media/astrapi69/backups/git/hub/astrapi69/";
-		String projectDirectoryName = targetProjectDirNamePrefix + targetProjectName;
-
-		File libsVersionsTomlFile = migrateToTomlVersions(gradleDirectory, targetProjectName,
-			targetProjectDirNamePrefix);
-
-		MigrationInfo migrationInfo = MigrationInfo.fromAbsolutePath(projectDirectoryName);
-		String libsVersionTomlMapAsString = newLibsVersionsTomlAsString(migrationInfo);
-		String libsVersionsTomlFileContent = ReadFileExtensions.fromFile(libsVersionsTomlFile);
-		assertEquals(libsVersionsTomlFileContent, libsVersionTomlMapAsString);
-
-		String sourceProjectName = DependenciesInfo.JAVA_LIBRARY_TEMPLATE_NAME;
-		String sourceProjectDirNamePrefix = "/run/media/astrapi69/backups/git/hub/astrapi69/";
-		GradleRunConfigurationsCopier.copyOnlyRunConfigurations(sourceProjectName,
-			targetProjectName, sourceProjectDirNamePrefix, targetProjectDirNamePrefix);
-	}
-
-	@Test
-	@Disabled
-	public void testMigrateToNewProjectStructure() throws IOException
-	{
-		File gradleDirectory;
-		String sourceProjectDirNamePrefix;
-		String targetProjectDirNamePrefix;
-		String sourceProjectName;
-		String targetProjectName;
-
-		gradleDirectory = getGradleDirectory();
-
-		targetProjectName = "throwable";
-		targetProjectDirNamePrefix = "/run/media/astrapi69/backups/git/hub/astrapi69/";
-		sourceProjectName = DependenciesInfo.JAVA_LIBRARY_TEMPLATE_NAME;
-		sourceProjectDirNamePrefix = "/run/media/astrapi69/backups/git/hub/astrapi69/";
-		ProjectTomlStructureInfo projectTomlStructureInfo = ProjectTomlStructureInfo.builder()
-			.gradleDirectory(gradleDirectory).sourceProjectName(sourceProjectName)
-			.sourceProjectDirNamePrefix(sourceProjectDirNamePrefix)
-			.targetProjectName(targetProjectName)
-			.targetProjectDirNamePrefix(targetProjectDirNamePrefix).build();
-
-		migrateToNewProjectStructure(projectTomlStructureInfo);
-	}
-
 	public static void migrateToNewProjectStructure(
 		ProjectTomlStructureInfo projectTomlStructureInfo) throws IOException
 	{
@@ -168,18 +130,6 @@ public class MigrateToTomlVersions
 		File libsVersionsTomlFile = migrateToTomlVersions(gradleDirectory, targetProjectName,
 			targetProjectDirNamePrefix);
 
-		GradleRunConfigurationsCopier.copyOnlyRunConfigurations(sourceProjectName,
-			targetProjectName, sourceProjectDirNamePrefix, targetProjectDirNamePrefix);
-	}
-
-	@Test
-	@Disabled
-	public void testCopyRunConfigurationsOfVersionCatalogUpdate() throws IOException
-	{
-		String sourceProjectName = DependenciesInfo.JAVA_LIBRARY_TEMPLATE_NAME;
-		String targetProjectName = "json-extensions";
-		String sourceProjectDirNamePrefix = "/run/media/astrapi69/backups/git/hub/astrapi69/";
-		String targetProjectDirNamePrefix = "/run/media/astrapi69/backups/git/hub/astrapi69/";
 		GradleRunConfigurationsCopier.copyOnlyRunConfigurations(sourceProjectName,
 			targetProjectName, sourceProjectDirNamePrefix, targetProjectDirNamePrefix);
 	}